python-blessed-api-tui/requirements.txt
blessed
requests
pyyaml
